commi concepto
1123

+++ SISJORSAC/SISJORSAC/DATA/Modelo/DetalleGasto.cs
@@ -10,11 +10,10 @@ namespace SISJORSAC.DATA.Modelo
     {
         public int COD_DETALLE { get; set; }
         public Gasto GASTO { get; set; }
-
         public ConceptoGasto ConceptoGasto { get; set; }
         public int ITEM { get; set; }
         public int CANTIDAD { get; set; }
-        public int PRECIO { get; set; }
+        public double PRECIO { get; set; }
 
     }
 }
